requirements.txt
pillow~=11.1.0
pyinstaller~=6.13.0
PyMuPDF~=1.25.5
pandas~=2.2.0
openpyxl~=3.1.0
tqdm~=4.66.0
PyQt6~=6.8.0
